// src/frontend/src/backend.d.ts
export interface Subtask {
    id: bigint;
    text: string;
    completed: boolean;
}
export interface DeviceInfo {
    status: DeviceStatus;
    isConnected: boolean;
    deviceId: string;
    deviceName: string;
    lastSync: Time;
}
export interface LogEvent {
    code: string;
    level: string;
    message: string;
    deviceId: string;
    timestamp: Time;
}
export interface ToDo {
    id: bigint;
    title: string;
    timeSpent: bigint;
    owner: Principal;
    createdAt: bigint;
    completed: boolean;
    dueDate?: bigint;
    description: string;
    timerSeconds?: bigint;
    category: string;
    priority: Priority;
    subtasks: Array<Subtask>;
}
// ...
export enum Priority {
    low = "low",
    high = "high",
    medium = "medium"
}